added weka to deps, added new AlmostRandomAgent that eliminates infinite games, added cap counting to gui and the beginnings of enclosure test
