+++ Go/pom.xml
@@ -61,5 +61,10 @@
   		<artifactId>junit</artifactId>
   		<version>4.11</version>
   	</dependency>
+    <dependency>
+      <groupId>nz.ac.waikato.cms.weka</groupId>
+      <artifactId>weka-stable</artifactId>
+      <version>3.6.6</version>
+    </dependency>
   </dependencies>
 </project>
