+++ Go/src/config/AlmostRandomAgentConfiguration.java
@@ -1,21 +1,21 @@
 package config;
 
+import agent.AlmostRandomAgent;
 import agent.IAgent;
-import agent.RandomAgent;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
 /**
- * Created by greg on 4/25/15.
+ * Created by grubino on 4/28/15.
  */
 @Configuration
 @ComponentScan(value = {"gui"})
-public class RandomAgentConfiguration {
+public class AlmostRandomAgentConfiguration {
 
     @Bean
     public IAgent getAgent() {
-        return new RandomAgent();
+        return new AlmostRandomAgent();
     }
 
 }
